pyproject.toml
[project]
name = "simple-add"
version = "0.1.0"
description = "simple addition"
authors = [
    {name = "youssefheshamm",email = "youssef_hesham@aucegypt.edu"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "sqlalchemy (>=2.0.42,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
